[AI/basic enemy] Final cleanup before PR. The main function was moved to a test file, which is included for reference.
